Refactored LetHelper
